Handle whitelisted origins
When a page or URL is added, the origin is added to `origins` collection. Only these origins are whitelisted by CORS.

Done to try resolve the error when calling local server from HTTP public site (https://stackoverflow.com/questions/66534759/cors-error-on-request-to-localhost-dev-server-from-remote-site) - Does not fix it. Need further checks

This also enforces that the extension works on only the registered pages.